refactor suggestion for python folders
